Create BisonFunctions
This file contains functions to be called in other files. Download this file and place it into the folder you are running the other file from.
In order to use it "import BisonFunctions as BF" at the top of the file. When you want to use a function you just type: BF.Portfolio_return() or whatever function you wish to use.